created cron job and realtime data update to app ui
