+++ .env.example
@@ -12,3 +12,6 @@ DIRECT_URL="postgresql://postgres:password@db.xxxxx.supabase.co:5432/postgres"
 RESEND_API_KEY=re_xxxxxxxxxxxx
 RESEND_FROM_EMAIL=onboarding@resend.dev
 NEXT_PUBLIC_APP_URL=http://localhost:3000
+
+# Cron (automated score sync via pg_cron)
+CRON_SECRET=
