+++ packages/api/src/index.ts
@@ -17,3 +17,4 @@ export type AppRouter = typeof appRouter;
 
 export { createContext } from "./context";
 export type { UserContext, CreateContextOptions, Context } from "./context";
+export { autoAdvanceScheduledTournaments } from "./routers/tournament";
